requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-timeout==3.0.1
attrs==21.2.0
cchardet==2.1.7
chardet==4.0.0
discord.py==1.7.3
greenlet==1.1.1
idna==3.2
multidict==5.1.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.2
yarl==1.6.3
